Update index files to be route.js
Remove unnecessary index.js files
