+++ src/server/admin/routes.js
@@ -1,7 +1,7 @@
-import { adminUsers } from '~/src/server/admin/users/index.js'
-import { adminTeams } from '~/src/server/admin/teams/index.js'
-import { adminFeatures } from '~/src/server/admin/features/index.js'
-import { adminPermissions } from '~/src/server/admin/permissions/index.js'
+import { adminUsers } from '~/src/server/admin/users/routes.js'
+import { adminTeams } from '~/src/server/admin/teams/routes.js'
+import { adminFeatures } from '~/src/server/admin/features/routes.js'
+import { adminPermissions } from '~/src/server/admin/permissions/routes.js'
 
 const admin = {
   plugin: {
