--- src/server/create/controllers/index.js
@@ -1,11 +1,0 @@
-import { isNameAvailableController } from '~/src/server/create/controllers/is-name-available.js'
-import { startController } from '~/src/server/create/controllers/start.js'
-import { chooseKindController } from '~/src/server/create/controllers/choose-kind.js'
-import { chooseKindFormController } from '~/src/server/create/controllers/choose-kind-form.js'
-
-export {
-  isNameAvailableController,
-  startController,
-  chooseKindController,
-  chooseKindFormController
-}
