--- src/server/create/journey-test-suite/controllers/index.js
@@ -1,11 +1,0 @@
-import { testSuiteDetailController } from '~/src/server/create/journey-test-suite/controllers/detail.js'
-import { testSuiteDetailFormController } from '~/src/server/create/journey-test-suite/controllers/detail-form.js'
-import { testSuiteSummaryController } from '~/src/server/create/journey-test-suite/controllers/summary.js'
-import { testSuiteCreateController } from '~/src/server/create/journey-test-suite/controllers/create.js'
-
-export {
-  testSuiteDetailController,
-  testSuiteDetailFormController,
-  testSuiteSummaryController,
-  testSuiteCreateController
-}
