--- src/server/create/microservice/controllers/index.js
@@ -1,11 +1,0 @@
-import { microserviceDetailController } from '~/src/server/create/microservice/controllers/detail.js'
-import { microserviceDetailFormController } from '~/src/server/create/microservice/controllers/detail-form.js'
-import { microserviceSummaryController } from '~/src/server/create/microservice/controllers/summary.js'
-import { microserviceCreateController } from '~/src/server/create/microservice/controllers/create.js'
-
-export {
-  microserviceDetailController,
-  microserviceDetailFormController,
-  microserviceSummaryController,
-  microserviceCreateController
-}
