+++ src/server/create/microservice/routes.js
@@ -1,9 +1,7 @@
-import {
-  microserviceCreateController,
-  microserviceDetailController,
-  microserviceDetailFormController,
-  microserviceSummaryController
-} from '~/src/server/create/microservice/controllers/index.js'
+import { microserviceDetailController } from '~/src/server/create/microservice/controllers/detail.js'
+import { microserviceDetailFormController } from '~/src/server/create/microservice/controllers/detail-form.js'
+import { microserviceSummaryController } from '~/src/server/create/microservice/controllers/summary.js'
+import { microserviceCreateController } from '~/src/server/create/microservice/controllers/create.js'
 
 const createMicroserviceRoutes = [
   {
