--- src/server/create/perf-test-suite/controllers/index.js
@@ -1,11 +1,0 @@
-import { perfTestSuiteDetailController } from '~/src/server/create/perf-test-suite/controllers/detail.js'
-import { perfTestSuiteDetailFormController } from '~/src/server/create/perf-test-suite/controllers/detail-form.js'
-import { perfTestSuiteSummaryController } from '~/src/server/create/perf-test-suite/controllers/summary.js'
-import { perfTestSuiteCreateController } from '~/src/server/create/perf-test-suite/controllers/create.js'
-
-export {
-  perfTestSuiteDetailController,
-  perfTestSuiteDetailFormController,
-  perfTestSuiteSummaryController,
-  perfTestSuiteCreateController
-}
