--- src/server/create/repository/controllers/index.js
@@ -1,13 +1,0 @@
-import { repositoryDetailController } from '~/src/server/create/repository/controllers/detail.js'
-import { repositoryDetailFormController } from '~/src/server/create/repository/controllers/detail-form.js'
-import { repositorySummaryController } from '~/src/server/create/repository/controllers/summary.js'
-import { repositoryCreateController } from '~/src/server/create/repository/controllers/create.js'
-import { repositorySuccessController } from '~/src/server/create/repository/controllers/success.js'
-
-export {
-  repositoryDetailController,
-  repositoryDetailFormController,
-  repositorySummaryController,
-  repositoryCreateController,
-  repositorySuccessController
-}
