+++ src/server/create/repository/routes.js
@@ -1,10 +1,8 @@
-import {
-  repositoryDetailFormController,
-  repositoryDetailController,
-  repositorySummaryController,
-  repositoryCreateController,
-  repositorySuccessController
-} from '~/src/server/create/repository/controllers/index.js'
+import { repositoryDetailController } from '~/src/server/create/repository/controllers/detail.js'
+import { repositoryDetailFormController } from '~/src/server/create/repository/controllers/detail-form.js'
+import { repositorySummaryController } from '~/src/server/create/repository/controllers/summary.js'
+import { repositoryCreateController } from '~/src/server/create/repository/controllers/create.js'
+import { repositorySuccessController } from '~/src/server/create/repository/controllers/success.js'
 
 const createRepositoryRoutes = [
   {
