--- src/server/deploy-service/controllers/deploy/index.js
@@ -1,17 +1,0 @@
-import { startDeployServiceController } from '~/src/server/deploy-service/controllers/deploy/start-deploy-service.js'
-import { optionsFormController } from '~/src/server/deploy-service/controllers/deploy/options-form.js'
-import { optionsController } from '~/src/server/deploy-service/controllers/deploy/options.js'
-import { detailsController } from '~/src/server/deploy-service/controllers/deploy/details.js'
-import { detailsFormController } from '~/src/server/deploy-service/controllers/deploy/details-form.js'
-import { summaryController } from '~/src/server/deploy-service/controllers/deploy/summary.js'
-import { deployController } from '~/src/server/deploy-service/controllers/deploy/deploy.js'
-
-export {
-  startDeployServiceController,
-  detailsController,
-  detailsFormController,
-  optionsController,
-  optionsFormController,
-  summaryController,
-  deployController
-}
