--- src/server/deploy-service/controllers/index.js
@@ -1,23 +1,0 @@
-import { availableVersionsController } from '~/src/server/deploy-service/controllers/available-versions.js'
-import { availableMemoryController } from '~/src/server/deploy-service/controllers/available-memory.js'
-import {
-  startDeployServiceController,
-  detailsController,
-  detailsFormController,
-  optionsController,
-  optionsFormController,
-  summaryController,
-  deployController
-} from '~/src/server/deploy-service/controllers/deploy/index.js'
-
-export {
-  availableVersionsController,
-  availableMemoryController,
-  startDeployServiceController,
-  detailsController,
-  detailsFormController,
-  optionsController,
-  optionsFormController,
-  summaryController,
-  deployController
-}
