--- src/server/deployments/controllers/index.js
@@ -1,4 +1,0 @@
-import { deploymentController } from '~/src/server/deployments/controllers/deployment.js'
-import { deploymentsListController } from '~/src/server/deployments/controllers/deployments-list.js'
-
-export { deploymentController, deploymentsListController }
