--- src/server/help/controllers/index.js
@@ -1,4 +1,0 @@
-import { cookiePolicy } from '~/src/server/help/controllers/cookie-policy.js'
-import { accessibilityStatement } from '~/src/server/help/controllers/accessibility-statement.js'
-
-export { accessibilityStatement, cookiePolicy }
