+++ src/server/router.js
@@ -1,21 +1,21 @@
 import inert from '@hapi/inert'
 
-import { admin } from '~/src/server/admin/index.js'
+import { admin } from '~/src/server/admin/routes.js'
 import { auth } from '~/src/server/auth/index.js'
-import { create } from '~/src/server/create/index.js'
-import { deployService } from '~/src/server/deploy-service/index.js'
-import { deployments } from '~/src/server/deployments/index.js'
-import { documentation } from '~/src/server/documentation/index.js'
-import { health } from '~/src/server/health/index.js'
-import { help } from '~/src/server/help/index.js'
-import { home } from '~/src/server/home/index.js'
-import { login } from '~/src/server/login/index.js'
-import { logout } from '~/src/server/logout/index.js'
-import { runningServices } from '~/src/server/running-services/index.js'
+import { create } from '~/src/server/create/routes.js'
+import { deployService } from '~/src/server/deploy-service/routes.js'
+import { deployments } from '~/src/server/deployments/routes.js'
+import { documentation } from '~/src/server/documentation/routes.js'
+import { health } from '~/src/server/health/routes.js'
+import { help } from '~/src/server/help/routes.js'
+import { home } from '~/src/server/home/routes.js'
+import { login } from '~/src/server/login/routes.js'
+import { logout } from '~/src/server/logout/routes.js'
+import { runningServices } from '~/src/server/running-services/routes.js'
 import { serveStaticFiles } from '~/src/server/common/helpers/serve-static-files.js'
-import { services } from '~/src/server/services/index.js'
-import { teams } from '~/src/server/teams/index.js'
-import { testSuites } from '~/src/server/test-suites/index.js'
+import { services } from '~/src/server/services/routes.js'
+import { teams } from '~/src/server/teams/routes.js'
+import { testSuites } from '~/src/server/test-suites/routes.js'
 
 const router = {
   plugin: {
