--- src/server/running-services/controllers/index.js
@@ -1,3 +1,0 @@
-import { runningServicesListController } from '~/src/server/running-services/controllers/running-services-list.js'
-
-export { runningServicesListController }
