--- src/server/services/about/controllers/index.js
@@ -1,8 +1,0 @@
-import { serviceController } from '~/src/server/services/about/controllers/service.js'
-import { serviceListController } from '~/src/server/services/about/controllers/service-list.js'
-import { serviceCreateStatusController } from '~/src/server/services/about/controllers/service-create-status.js'
-export {
-  serviceController,
-  serviceListController,
-  serviceCreateStatusController
-}
