+++ src/server/services/about/routes.js
@@ -1,8 +1,6 @@
-import {
-  serviceController,
-  serviceListController,
-  serviceCreateStatusController
-} from '~/src/server/services/about/controllers/index.js'
+import { serviceController } from '~/src/server/services/about/controllers/service.js'
+import { serviceListController } from '~/src/server/services/about/controllers/service-list.js'
+import { serviceCreateStatusController } from '~/src/server/services/about/controllers/service-create-status.js'
 import { provideTabs } from '~/src/server/services/helpers/provide-tabs.js'
 import { provideService } from '~/src/server/services/helpers/provide-service.js'
