--- src/server/services/secrets/controllers/index.js
@@ -1,13 +1,0 @@
-import { allSecretsController } from '~/src/server/services/secrets/controllers/all.js'
-import { environmentSecretsController } from '~/src/server/services/secrets/controllers/environment.js'
-import { updateSecretController } from '~/src/server/services/secrets/controllers/update.js'
-import { createSecretController } from '~/src/server/services/secrets/controllers/create.js'
-import { updateSecretFormController } from '~/src/server/services/secrets/controllers/update-form.js'
-
-export {
-  allSecretsController,
-  environmentSecretsController,
-  updateSecretController,
-  createSecretController,
-  updateSecretFormController
-}
