+++ src/server/services/secrets/routes.js
@@ -1,10 +1,8 @@
-import {
-  allSecretsController,
-  environmentSecretsController,
-  updateSecretController,
-  createSecretController,
-  updateSecretFormController
-} from '~/src/server/services/secrets/controllers/index.js'
+import { allSecretsController } from '~/src/server/services/secrets/controllers/all.js'
+import { environmentSecretsController } from '~/src/server/services/secrets/controllers/environment.js'
+import { updateSecretController } from '~/src/server/services/secrets/controllers/update.js'
+import { createSecretController } from '~/src/server/services/secrets/controllers/create.js'
+import { updateSecretFormController } from '~/src/server/services/secrets/controllers/update-form.js'
 import { scopes } from '~/src/server/common/constants/scopes.js'
 import { authScope } from '~/src/server/common/helpers/auth/auth-scope.js'
 import { provideTabs } from '~/src/server/services/helpers/provide-tabs.js'
