--- src/server/services/terminal/controllers/index.js
@@ -1,9 +1,0 @@
-import { launchTerminalController } from '~/src/server/services/terminal/controllers/launch-terminal.js'
-import { terminalController } from '~/src/server/services/terminal/controllers/terminal.js'
-import { terminalBrowserController } from '~/src/server/services/terminal/controllers/terminal-browser.js'
-
-export {
-  launchTerminalController,
-  terminalController,
-  terminalBrowserController
-}
