--- src/server/teams/controllers/add/index.js
@@ -1,4 +1,0 @@
-import { addMemberFormController } from '~/src/server/teams/controllers/add/member-form.js'
-import { addMemberController } from '~/src/server/teams/controllers/add/member.js'
-
-export { addMemberFormController, addMemberController }
