--- src/server/teams/controllers/edit/index.js
@@ -1,9 +1,0 @@
-import { teamEditStartController } from '~/src/server/teams/controllers/edit/start.js'
-import { teamDetailsFormController } from '~/src/server/teams/controllers/edit/team-details-form.js'
-import { teamDetailsController } from '~/src/server/teams/controllers/edit/team-details.js'
-
-export {
-  teamEditStartController,
-  teamDetailsFormController,
-  teamDetailsController
-}
