+++ src/server/teams/routes.js
@@ -1,13 +1,11 @@
-import {
-  teamController,
-  teamsListController,
-  teamEditStartController,
-  teamDetailsFormController,
-  teamDetailsController,
-  addMemberController,
-  addMemberFormController,
-  removeMemberController
-} from '~/src/server/teams/controllers/index.js'
+import { addMemberFormController } from '~/src/server/teams/controllers/add/member-form.js'
+import { addMemberController } from '~/src/server/teams/controllers/add/member.js'
+import { teamEditStartController } from '~/src/server/teams/controllers/edit/start.js'
+import { teamDetailsFormController } from '~/src/server/teams/controllers/edit/team-details-form.js'
+import { teamDetailsController } from '~/src/server/teams/controllers/edit/team-details.js'
+import { teamController } from '~/src/server/teams/controllers/team.js'
+import { teamsListController } from '~/src/server/teams/controllers/teams-list.js'
+import { removeMemberController } from '~/src/server/teams/controllers/remove/member.js'
 import { authScope } from '~/src/server/common/helpers/auth/auth-scope.js'
 import { sessionNames } from '~/src/server/common/constants/session-names.js'
 import { provideFormContextValues } from '~/src/server/common/helpers/form/provide-form-context-values.js'
