--- src/server/test-suites/controllers/index.js
@@ -1,15 +1,0 @@
-import { testSuiteController } from '~/src/server/test-suites/controllers/test-suite.js'
-import { testSuiteListController } from '~/src/server/test-suites/controllers/test-suite-list.js'
-import { testSuiteStatusController } from '~/src/server/test-suites/controllers/test-suite-status.js'
-import { triggerTestSuiteRunController } from '~/src/server/test-suites/controllers/trigger-test-suite-run.js'
-import { testResultsController } from '~/src/server/test-suites/controllers/test-results.js'
-import { testSuiteReportController } from '~/src/server/test-suites/controllers/test-suite-report.js'
-
-export {
-  testSuiteController,
-  testSuiteListController,
-  testSuiteStatusController,
-  triggerTestSuiteRunController,
-  testResultsController,
-  testSuiteReportController
-}
