+++ src/server/test-suites/routes.js
@@ -1,12 +1,10 @@
 import { provideFormContextValues } from '~/src/server/common/helpers/form/provide-form-context-values.js'
-import {
-  testSuiteController,
-  testSuiteListController,
-  testSuiteStatusController,
-  triggerTestSuiteRunController,
-  testResultsController,
-  testSuiteReportController
-} from '~/src/server/test-suites/controllers/index.js'
+import { testSuiteController } from '~/src/server/test-suites/controllers/test-suite.js'
+import { testSuiteListController } from '~/src/server/test-suites/controllers/test-suite-list.js'
+import { testSuiteStatusController } from '~/src/server/test-suites/controllers/test-suite-status.js'
+import { triggerTestSuiteRunController } from '~/src/server/test-suites/controllers/trigger-test-suite-run.js'
+import { testResultsController } from '~/src/server/test-suites/controllers/test-results.js'
+import { testSuiteReportController } from '~/src/server/test-suites/controllers/test-suite-report.js'
 import { stopTestSuiteController } from '~/src/server/test-suites/controllers/stop-test-suite.js'
 
 const testSuites = {
